Update Front End User

+++ database/migrations/2021_10_21_091331_create_discounts_table.php
@@ -15,7 +15,7 @@ class CreateDiscountsTable extends Migration
     {
         Schema::create('discounts', function (Blueprint $table) {
             $table->increments('id');
-            $table->decimal('amounts');
+            $table->double('amounts');
             $table->dateTime('start_date');
             $table->dateTime('expired_date');
             $table->string('description');
